test: add chord diagram rendering test
